requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
openai
upstash-redis
